fix: migrate Timefold 1.x to 2.x and fix Spring Boot 3.x compatibility
- Upgrade timefold-solver 1.x -> 2.1.0 (bluetape4k-dependencies 1.1.3 BOM)
- Update artifact: timefold-solver-persistence-common removed in 2.x
- Update artifact: timefold-solver-test removed; ConstraintVerifier moved to core
- Fix score package: buildin.*.*Score -> HardMediumSoftScore/HardSoftScore (top-level)
- Remove Long score types (BendableLong/HardSoftLong/HardMediumSoftLong/SimpleLong) removed in 2.x
- Fix score factory params: Int -> Long in Timefold 2.x
- Fix PlanningId import: domain.lookup -> domain.common
- Fix SolverManager: 2 type params -> 1 type param in 2.x
- Fix SolverJobBuilder: withBestSolutionEventConsumer receives NewBestSolutionEvent
- Fix BedPlan.score: non-null -> nullable (HardMediumSoftScore? = null)
- Remove GraalVM native plugin from quickstart modules (timefold-solver-spring-boot-starter
  2.x compiled against Spring 7.x/AOT APIs incompatible with Spring Boot 3.x/Spring 6.x)
- Disable analyze test: SolutionManager.analyze() is Enterprise-only in Timefold 2.x

+++ exposed/jdbc-examples/src/test/kotlin/timefold/workshop/persistence/exposed/HardMediumSoftScoreTest.kt
@@ -1,6 +1,6 @@
 package timefold.workshop.persistence.exposed
 
-import ai.timefold.solver.core.api.score.buildin.hardmediumsoft.HardMediumSoftScore
+import ai.timefold.solver.core.api.score.HardMediumSoftScore
 import io.bluetape4k.exposed.dao.entityToStringBuilder
 import io.bluetape4k.exposed.dao.idEquals
 import io.bluetape4k.exposed.dao.idHashCode
@@ -51,9 +51,9 @@ class HardMediumSoftScoreTest: AbstractScoreExposedTest() {
 
             val name = faker.name().name()
             val hardMediumSoftScore = HardMediumSoftScore.of(
-                faker.random().nextInt(),
-                faker.random().nextInt(),
-                faker.random().nextInt()
+                faker.random().nextLong(),
+                faker.random().nextLong(),
+                faker.random().nextLong()
             )
 
             val id = T1.insertAndGetId {
@@ -80,9 +80,9 @@ class HardMediumSoftScoreTest: AbstractScoreExposedTest() {
             val saved = E1.new {
                 this.name = faker.name().name()
                 this.hardMediumSoftScore = HardMediumSoftScore.of(
-                    faker.random().nextInt(),
-                    faker.random().nextInt(),
-                    faker.random().nextInt()
+                    faker.random().nextLong(),
+                    faker.random().nextLong(),
+                    faker.random().nextLong()
                 )
             }
             log.debug { "saved=$saved" }
